2. TESTE DE TRANSFORMAÇÃO DE DADOS/data_transformation_python/requirements.txt
pandas
tabula-py
jpype1
packaging
